refactor: move more elements
